Make Compatible with Eclipse 3.6
I suspected that the plug-in would be compatible with 3.6 since there should be minor changes between 3.6 and 3.7. The purpose of this is so that it can be used with Helios (3.6), and thereby be installable from the marketplace with IBM Rational Application Developer Version 8.5.x which is based on Helios.

IBM's JDK does not play nice with JMockit, and needs the help of this plug-in for the tests to run from the IDE. I have tested the above changes with stand-alone Helios 3.6 and with IBM Rational Application Developer 8.5.5.1. Both worked without any problems. So, it seems that the code-base is perfectly backward compatible with Helios.